fix: orginizing folder structure
fix: removing model-store to simplify code

fix: improving components
