+++ .env
@@ -1,5 +1,4 @@
 NODE_ENV='production'
-PORT=''
 ACCESS_TOKEN_COOKIE_NAME=''
 BASE_URL=''
 CSRF_TOKEN_API_PATH=''
@@ -14,4 +13,4 @@ SITE_NAME=''
 USER_INFO_COOKIE_NAME=''
 APP_ID=''
 MFE_CONFIG_API_URL=''
-LEARNING_MFE_BASE_URL=''
+LEARNING_BASE_URL=''
