+++ .env.development
@@ -14,4 +14,4 @@ SITE_NAME=localhost
 USER_INFO_COOKIE_NAME='edx-user-info'
 APP_ID=''
 MFE_CONFIG_API_URL=''
-LEARNING_MFE_BASE_URL='http://localhost:2000/'
+LEARNING_BASE_URL='http://localhost:2000/'
