--- docs/how_tos/i18n.rst
@@ -1,5 +1,0 @@
-####################
-React App i18n HOWTO
-####################
-
-This document has moved to the frontend-platform repo: https://github.com/openedx/frontend-platform/blob/master/docs/how_tos/i18n.rst
