+++ package.json
@@ -36,25 +36,19 @@
   },
   "dependencies": {
     "@edx/brand": "npm:@edx/brand-openedx@1.1.0",
-    "@edx/frontend-lib-special-exams": "2.1.0",
     "@edx/frontend-platform": "2.6.2",
     "@edx/paragon": "19.25.3",
-    "@fortawesome/fontawesome-svg-core": "1.2.36",
-    "@fortawesome/free-brands-svg-icons": "5.15.4",
     "@fortawesome/free-regular-svg-icons": "5.15.4",
     "@fortawesome/free-solid-svg-icons": "5.15.4",
     "@fortawesome/react-fontawesome": "0.2.0",
     "@reduxjs/toolkit": "1.8.1",
-    "core-js": "3.25.5",
     "prop-types": "15.8.1",
     "react": "16.14.0",
     "react-dom": "16.14.0",
-    "react-helmet": "6.1.0",
     "react-redux": "7.2.9",
     "react-router": "5.3.4",
     "react-router-dom": "5.3.4",
-    "redux": "4.2.0",
-    "regenerator-runtime": "0.13.10"
+    "redux": "4.2.0"
   },
   "devDependencies": {
     "@edx/frontend-build": "12.3.0",
