--- src/features/course-home/data/index.js
@@ -1,5 +1,0 @@
-export {
-  fetchCourseOutline,
-} from 'features/course-home/data/thunks';
-
-export { reducer } from 'features/course-home/data/slice';
