--- src/features/course-home/outline/index.js
@@ -1,2 +1,0 @@
-/* eslint-disable no-restricted-exports */
-export { default } from 'features/course-home/outline/CourseOutline';
