--- src/features/generic/model-store/index.js
@@ -1,5 +1,0 @@
-export {
-  useModel,
-  reducer,
-  addModel,
-} from 'features/generic/model-store/slice';
