+++ src/features/outline/data/index.js
@@ -1,0 +1,5 @@
+export {
+  fetchCourseOutline,
+} from 'features/outline/data/thunks';
+
+export { reducer } from 'features/outline/data/slice';
