+++ src/features/outline/index.js
@@ -1,0 +1,2 @@
+/* eslint-disable no-restricted-exports */
+export { default } from 'features/outline/CourseOutline';
