--- src/i18n/messages/ar.json
@@ -1,1 +1,0 @@
-{}
