--- src/i18n/messages/ca.json
@@ -1,1 +1,0 @@
-{}
