--- src/i18n/messages/es_419.json
@@ -1,1 +1,0 @@
-{}
