--- src/i18n/messages/fr.json
@@ -1,1 +1,0 @@
-{}
