--- src/i18n/messages/he.json
@@ -1,1 +1,0 @@
-{}
