--- src/i18n/messages/id.json
@@ -1,1 +1,0 @@
-{}
