--- src/i18n/messages/ko_kr.json
@@ -1,1 +1,0 @@
-{}
