--- src/i18n/messages/pl.json
@@ -1,1 +1,0 @@
-{}
