--- src/i18n/messages/pt_br.json
@@ -1,1 +1,0 @@
-{}
