--- src/i18n/messages/ru.json
@@ -1,1 +1,0 @@
-{}
