--- src/i18n/messages/th.json
@@ -1,1 +1,0 @@
-{}
