--- src/i18n/messages/uk.json
@@ -1,1 +1,0 @@
-{}
