--- src/i18n/messages/zh_CN.json
@@ -1,1 +1,0 @@
-{}
