+++ src/setupTest.js
@@ -1,2 +1,1 @@
-import 'core-js/stable';
 import 'regenerator-runtime/runtime';
