+++ src/store.js
@@ -1,12 +1,10 @@
 import { configureStore } from '@reduxjs/toolkit';
-import { reducer as courseHomeReducer } from 'features/course-home/data';
-import { reducer as modelsReducer } from 'features/generic/model-store';
+import { reducer as outlineReducer } from 'features/outline/data';
 
 export default function initializeStore() {
   return configureStore({
     reducer: {
-      models: modelsReducer,
-      courseHome: courseHomeReducer,
+      outline: outlineReducer,
     },
   });
 }
